Move read_from and write_to operations to lib.rs
